requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1

langchain>=0.2.14
langchain-community>=0.2.11
langgraph>=0.2.30
chromadb>=0.5.3

sqlalchemy>=2.0.30
pandas>=2.2.2
tiktoken>=0.7.0
cohere>=5.5.6
langchain-cohere>=0.2.0

streamlit==1.39.0
pandas==2.2.3
requests==2.32.3